+++ src/common/global_types.rs
@@ -1,14 +1,14 @@
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
-pub type Pressure = f64;
-pub type DepthType = f64;
+pub type Pressure = f32;
+pub type DepthType = f32;
 pub type GradientFactor = u8;
 pub type GradientFactors = (u8, u8);
 pub type MbarPressure = i32;
-pub type AscentRatePerMinute = f64;
-pub type Cns = f64;
-pub type Otu = f64;
+pub type AscentRatePerMinute = f32;
+pub type Cns = f32;
+pub type Otu = f32;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
